// src/types/sign.ts
export interface LoginStatus {
  email: string
  password: string
}
export interface SignUpStatus {
  email: string
  password: string
  phoneNumber: string
  fullName: string
  profilePhoto: string
  businessName: string
  businessType: string
  description: string
  address: string
  city: string
  latitude: number
  longitude: number
  phone: string
}